+++ packages/eslint-plugin/rules/__tests__/components-no-missing-40px-size-prop.js
@@ -71,20 +71,6 @@ ruleTester.run( 'components-no-missing-40px-size-prop', rule, {
 				<WPButton __next40pxDefaultSize />
 			`,
 		},
-		// FormFileUpload with render prop (special case)
-		{
-			code: `
-				import { FormFileUpload } from '@wordpress/components';
-				<FormFileUpload render={({ open }) => <button onClick={open}>Upload</button>} />
-			`,
-		},
-		// FormFileUpload with __next40pxDefaultSize
-		{
-			code: `
-				import { FormFileUpload } from '@wordpress/components';
-				<FormFileUpload __next40pxDefaultSize />
-			`,
-		},
 		// Component with dynamic size prop (assumes it could be non-default)
 		{
 			code: `
@@ -110,17 +96,9 @@ ruleTester.run( 'components-no-missing-40px-size-prop', rule, {
 		{
 			code: `
 				import {
-					BorderBoxControl,
-					BorderControl,
-					ComboboxControl,
 					CustomSelectControl,
-					FontAppearanceControl,
-					FontFamilyControl,
-					FontSizePicker,
 					FormTokenField,
 					InputControl,
-					LetterSpacingControl,
-					LineHeightControl,
 					NumberControl,
 					RangeControl,
 					SelectControl,
@@ -128,21 +106,13 @@ ruleTester.run( 'components-no-missing-40px-size-prop', rule, {
 					UnitControl,
 				} from '@wordpress/components';
 				<>
-					<BorderBoxControl __next40pxDefaultSize />
-					<BorderControl __next40pxDefaultSize />
-					<ComboboxControl __next40pxDefaultSize />
 					<CustomSelectControl __next40pxDefaultSize />
-					<FontAppearanceControl __next40pxDefaultSize />
-					<FontFamilyControl __next40pxDefaultSize />
-					<FontSizePicker __next40pxDefaultSize />
 					<FormTokenField __next40pxDefaultSize />
 					<InputControl __next40pxDefaultSize />
-					<LetterSpacingControl __next40pxDefaultSize />
-					<LineHeightControl __next40pxDefaultSize />
 					<NumberControl __next40pxDefaultSize />
-					<RangeControl __next40pxDefaultSize />
+					<RangeControl />
 					<SelectControl __next40pxDefaultSize />
-					<ToggleGroupControl __next40pxDefaultSize />
+					<ToggleGroupControl />
 					<UnitControl __next40pxDefaultSize />
 				</>
 			`,
@@ -211,18 +181,6 @@ ruleTester.run( 'components-no-missing-40px-size-prop', rule, {
 				{
 					messageId: 'missingProp',
 					data: { component: 'InputControl' },
-				},
-			],
-		},
-		// FormFileUpload without __next40pxDefaultSize or render
-		{
-			code: `
-				import { FormFileUpload } from '@wordpress/components';
-				<FormFileUpload onChange={handleChange} />
-			`,
-			errors: [
-				{
-					messageId: 'missingPropFormFileUpload',
 				},
 			],
 		},
@@ -336,14 +294,6 @@ ruleTester.run(
 				import InputControl from '../input-control';
 				<InputControl />
 			`,
-			},
-			// FormFileUpload relative import with render prop
-			{
-				code: `
-				import { FormFileUpload } from '../form-file-upload';
-				<FormFileUpload render={({ open }) => <button onClick={open}>Upload</button>} />
-			`,
-				options: [ { checkLocalImports: true } ],
 			},
 		],
 		invalid: [],